Renaming and reviewing quickly app.channel.controllers.py
